Fix AVR build warnings and linking errors (#6)
* Fix all AVR compilation warnings and errors

- Fix signed/unsigned comparison in LD2420.cpp:164
- Fix constructor member initialization order in LD245X.hpp
- Fix signed/unsigned comparison in Debug.hpp macros
- Fix format specifier for uint32_t (use %lu)
- Initialize val in CircularBuffer.hpp to avoid uninitialized warning
- Remove redundant HardwareSerial/SoftwareSerial includes from examples

Agent-Logs-Url: https://github.com/Fiooodooor/HLK-LD245X/sessions/fcd3163a-cb99-4369-95ad-07fcfe4cda6e


* Fix AVR linking error - use HardwareSerial on AVR

Added ARDUINO_ARCH_AVR to the conditional check for SERIAL_TYPE
to ensure AVR platforms use HardwareSerial instead of SoftwareSerial.
Removed unnecessary FORCE_HARDWARE_SERIAL defines from examples.

Agent-Logs-Url: https://github.com/Fiooodooor/HLK-LD245X/sessions/fcd3163a-cb99-4369-95ad-07fcfe4cda6e


---------

+++ src/LD2420.cpp
@@ -161,7 +161,7 @@ bool LD2420::queryFirmwareVersion()
     if (retCmd != 0x0100 || retVal != 0) return false;
 
     uint16_t strLen = word(frameBuffer[5], frameBuffer[4]);
-    if (frameBufferBytesRead < 6 + strLen) return false;
+    if (frameBufferBytesRead < static_cast<int>(6 + strLen)) return false;
 
     snprintf(ld2420_firmware_string, sizeof(ld2420_firmware_string),
              "%.*s", strLen, reinterpret_cast<const char*>(frameBuffer + 6));
